refactor data_generator
