+++ CMakeLists.txt
@@ -505,7 +505,7 @@ add_executable(agora
 target_link_libraries(agora ${COMMON_LIBS})
 
 add_executable(data_generator
-  ${CMAKE_CURRENT_SOURCE_DIR}/src/data_generator/data_generator_main.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/data_generator/test_vector_generator.cc
   $<TARGET_OBJECTS:common_sources_lib>)
 target_link_libraries(data_generator ${COMMON_LIBS})
 target_compile_definitions(data_generator PRIVATE GENERATE_DATA)
